refactored custom port painter functions into new file, ports now have the correct shape according to their data type after loading a pipeline graph
